Refactor data types and update EF Core model
Updated various property data types from int to string and decimal to double across multiple classes. Enhanced several entities to inherit from BaseEntity<string> and added auditing properties. Renamed AgencyLocation table to Locations and updated foreign key relationships. Introduced new auto-generated entity type classes for EF Core model, including configurations for properties, keys, indexes, and foreign keys. Updated migration files and DbContext to reflect these changes. Adjusted project settings in Turo.csproj and modified Program.cs for culture settings.

+++ Turo.Domain/Entities/Agency.cs
@@ -22,7 +22,7 @@ namespace Turo.Domain.Entities
         public new ICollection<Car>? CarsOwned { get; set; }  // List of cars owned by the agency
 
         // Pricing & Discounts
-        public decimal DailyRentalPrice { get; set; }  // Average daily rental price
+        public double DailyRentalPrice { get; set; }  // Average daily rental price
         public string? PricingStrategy { get; set; }  // e.g., "Dynamic pricing" or "Fixed pricing"
         public ICollection<DiscountOffer>? DiscountOffers { get; set; }  // List of discounts/offers provided by the agency
